+++ crates/wdl-lint/src/rules/meta_description.rs
@@ -81,6 +81,7 @@ impl Rule for MetaDescriptionRule {
     fn related_rules(&self) -> &[&'static str] {
         &[
             "ParameterMetaMatched",
+            "ParameterDescription",
             "OutputSection",
             "RequirementsSection",
             "RuntimeSection",